back-end/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10 
dj-database-url==0.5.0     
Django==3.2.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
idna==3.3
Pillow==8.4.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
Unidecode==1.3.2
urllib3==1.26.8